sqlite [nfc]: Move sqlite tests to their own file
