+++ src/storage/__tests__/sqlite-test.js
@@ -1,0 +1,32 @@
+// @flow strict-local
+import { openDatabase } from 'expo-sqlite';
+// $FlowFixMe[untyped-import]
+import sqlite3 from 'sqlite3';
+
+/* eslint-disable no-underscore-dangle */
+
+describe('sqlite3', () => {
+  test('smoke', async () => {
+    const db = new sqlite3.Database(':memory:');
+    const result = await new Promise((resolve, reject) => {
+      db.get('SELECT 42 as n', (err, row) => (err ? reject(err) : resolve(row)));
+    });
+    expect(result).toEqual({ n: 42 });
+  });
+});
+
+describe('expo-sqlite', () => {
+  test('smoke', async () => {
+    const db = openDatabase('test.db');
+    const result = await new Promise((resolve, reject) => {
+      db.readTransaction(
+        tx => {
+          tx.executeSql('SELECT 42 AS n', [], (t, r) => resolve(r));
+        },
+        reject,
+        resolve,
+      );
+    });
+    expect(result.rows._array).toEqual([{ n: 42 }]);
+  });
+});
